ingestion/pyproject.toml
[tool.poetry]
name = "dvp-ingestion"
version = "0.1.0"
description = "Data Vanguards Platform - Ingestion Service"
authors = ["Data Vanguards Team"]
readme = "README.md"
packages = [
    { include = "ingestion", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pydantic = "^2.5.3"
boto3 = "^1.34.15"  # For S3 support
aioboto3 = "^12.0.0"  # Updated to be compatible with boto3 1.34
aiohttp = "^3.9.0"  # For async HTTP requests
aiosignal = "^1.3.1"  # Required by aiohttp
async-timeout = "^4.0.3"  # Required by aiohttp
attrs = "^23.1.0"  # Required by aiohttp
frozenlist = "^1.4.0"  # Required by aiohttp
multidict = "^6.0.4"  # Required by aiohttp
yarl = "^1.9.3"  # Required by aiohttp
idna = "^3.4"  # Required by yarl
requests = "^2.31.0"  # For HTTP requests
requests-toolbelt = "^1.0.0"
clickhouse-connect = "^0.7.0"  # For ClickHouse metrics
psutil = "^5.9.0"  # For system metrics
dvp-observability = { file = "wheels/dvp_observability-0.3.0-py3-none-any.whl" }
structlog = "^23.1.0"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-instrumentation-aiohttp-client = "^0.41b0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
gql = "^3.5.0"
requests-oauthlib = "^1.3.1"
oauthlib = "^3.2.2"
moto = "^4.1.13"
pytest-mock = "^3.11.1"
dlt = "^0.4.5"
aws-xray-sdk = "^2.12.0"
clickhouse-driver = "^0.2.6"
pandas = "^2.1.0"
tabulate = "^0.9.0"
plotly = "^5.18.0"
rich = "^13.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
dvp-observability = { file = "wheels/dvp_observability-0.3.0-py3-none-any.whl" }
pytest-mock = "^3.14.0"
black = "^23.12.0"
isort = "^5.13.0"
mypy = "^1.8.0"
pylint = "^3.0.0"


[tool.poetry.scripts]
explore-nsw-surfing = "exploration.nsw_surfing_explorer:explore_nsw_surfing"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = [
    "src"
]
addopts = "-v --cov=ingestion --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]

[tool.python]
py_compile = false
